requirements.txt
streamlit
pymupdf
sentence-transformers
faiss-cpu
transformers
torch
